# lab3/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  serializer.UserJSON:
    properties:
      id:
        type: string
      is_moderator:
        type: boolean
      login:
        type: string
      password:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@bigorequest.com
    name: API Support
    url: http://localhost:8080
  description: API для управления расчётами времени и сложности Классов сложности
  license:
    name: MIT
  title: Big O Request API
  version: "1.0"
paths:
...
  /users/{login}/info:
    get:
      description: Возвращает данные профиля (доступен только тот, чей UUID совпадает
        с user_id в токене).
      parameters:
      - description: Логин пользователя
        in: path
        name: login
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Профиль пользователя
          schema:
            $ref: '#/definitions/serializer.UserJSON'
        "400":
          description: Проблема с авторизацией/получением user_id"}
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Пользователи не совпадают
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить профиль пользователя
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновляет профиль пользователя (может делать только сам пользователь).
      parameters:
      - description: Логин пользователя
        in: path
        name: login
        required: true
        type: string
      - description: Новые данные профиля
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/serializer.UserJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённый профиль
          schema:
            $ref: '#/definitions/serializer.UserJSON'
        "400":
          description: Ошибка запроса или авторизации
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещён
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить профиль пользователя
      tags:
      - users
  /users/signin:
    post:
      consumes:
      - application/json
      description: Принимает логин/пароль, возвращает jwt-токен в формате {"token":"..."}.
      parameters:
      - description: Логин и пароль
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/serializer.UserJSON'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход (получение токена)
      tags:
      - users
  /users/signout:
    post:
      description: Удаляет токен текущего пользователя из хранилища. Возвращает {"status":"signed_out"}.
      produces:
      - application/json
      responses:
        "200":
          description: status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Проблема с получением user_id
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка при удалении токена
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Выход (удаление токена)
      tags:
      - users
  /users/signup:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя. Возвращает URL созданного ресурса
        в Location и тело созданного пользователя.
      parameters:
      - description: Параметры нового пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/serializer.UserJSON'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь создан
          schema:
            $ref: '#/definitions/serializer.UserJSON'
        "400":
          description: Ошибка валидации или входных данных
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"